refactor operation handlers
