+++ x-pack/platform/plugins/shared/dashboard_agent/server/tools/manage_dashboard/operations.ts
@@ -6,27 +6,18 @@
  */
 
 import { z } from '@kbn/zod/v4';
-import { v4 as uuidv4 } from 'uuid';
 import { SupportedChartType } from '@kbn/agent-builder-common/tools/tool_result';
-import type {
-  AttachmentPanel,
-  DashboardAttachmentData,
-  DashboardSection,
-} from '@kbn/dashboard-agent-common';
+import type { AttachmentPanel, DashboardAttachmentData } from '@kbn/dashboard-agent-common';
 import { panelGridSchema, sectionGridSchema } from '@kbn/dashboard-agent-common';
 import type { Logger } from '@kbn/core/server';
-import { MARKDOWN_EMBEDDABLE_TYPE } from '@kbn/dashboard-markdown/server';
-import {
-  appendPanelsToDashboard,
-  findPanelById,
-  findSectionIndex,
-  getWidgetsBottomY,
-  removePanelsFromDashboard,
-  updatePanelInDashboard,
-} from './dashboard_state';
 import type { ResolveVisualizationConfig } from './inline_visualization';
-import { createVisualizationFailureResult } from './inline_visualization';
 import type { VisualizationFailure } from './utils';
+import { executeOperationHandler } from './operations/handlers';
+import type { OperationExecutionContext } from './operations/types';
+import {
+  collectVisualizationCreationRequests,
+  resolveVisualizationCreationRequests,
+} from './operations/visualization_creation';
 
 export const setMetadataOperationSchema = z.object({
   operation: z.literal('set_metadata'),
@@ -216,174 +207,8 @@ interface ExecuteDashboardOperationsParams {
   resolveVisualizationConfig?: ResolveVisualizationConfig;
 }
 
-type VisualizationPanelInput = z.infer<typeof visualizationPanelInputSchema>;
-type CreateVisualizationPanelInput = z.infer<typeof createVisualizationPanelSchema>;
-type ResolvedVisualizationPanel = Awaited<ReturnType<ResolveVisualizationConfig>>;
-type VisualizationCreationOperationType = 'add_section' | 'create_visualization_panels';
-type VisualizationCreationRequest =
-  | {
-      operationType: 'add_section';
-      panelInput: VisualizationPanelInput;
-    }
-  | {
-      operationType: 'create_visualization_panels';
-      panelInput: CreateVisualizationPanelInput;
-      sectionId?: string;
-    };
-interface ResolvedVisualizationCreationRequest {
-  request: VisualizationCreationRequest;
-  resolvedPanel: ResolvedVisualizationPanel;
-}
-
-/**
- * Collect inline visualization creation work by operation index so it can be
- * resolved up front in parallel and then applied later in original operation order.
- */
-const collectVisualizationCreationRequests = (
-  operations: DashboardOperation[]
-): Map<number, VisualizationCreationRequest[]> => {
-  const requestsByOperationIndex = new Map<number, VisualizationCreationRequest[]>();
-
-  for (const [operationIndex, operation] of operations.entries()) {
-    switch (operation.operation) {
-      case 'add_section': {
-        if (!operation.panels) {
-          break;
-        }
-
-        requestsByOperationIndex.set(
-          operationIndex,
-          operation.panels.map((panelInput) => ({
-            operationType: operation.operation,
-            panelInput,
-          }))
-        );
-        break;
-      }
-      case 'create_visualization_panels': {
-        requestsByOperationIndex.set(
-          operationIndex,
-          operation.panels.map((panelInput) => ({
-            operationType: operation.operation,
-            panelInput,
-            sectionId: panelInput.sectionId,
-          }))
-        );
-        break;
-      }
-      default:
-        break;
-    }
-  }
-
-  return requestsByOperationIndex;
-};
-
-/**
- * Resolve all collected inline visualization creation requests up front while
- * keeping results grouped by their source operation for ordered application later.
- */
-const resolveVisualizationCreationRequests = async ({
-  requestsByOperationIndex,
-  resolveVisualizationConfig,
-}: {
-  requestsByOperationIndex: Map<number, VisualizationCreationRequest[]>;
-  resolveVisualizationConfig?: ResolveVisualizationConfig;
-}): Promise<Map<number, ResolvedVisualizationCreationRequest[]>> => {
-  if (requestsByOperationIndex.size === 0) {
-    return new Map();
-  }
-
-  if (!resolveVisualizationConfig) {
-    throw new Error(
-      'Inline visualization resolver is required for visualization creation operations.'
-    );
-  }
-
-  const resolvedRequestsByOperationIndex = await Promise.all(
-    Array.from(requestsByOperationIndex.entries()).map(
-      async ([operationIndex, requests]): Promise<
-        readonly [number, ResolvedVisualizationCreationRequest[]]
-      > =>
-        [
-          operationIndex,
-          await Promise.all(
-            requests.map(async (request) => ({
-              request,
-              resolvedPanel: await resolveVisualizationConfig({
-                operationType: request.operationType,
-                identifier: request.panelInput.query,
-                nlQuery: request.panelInput.query,
-                index: request.panelInput.index,
-                chartType: request.panelInput.chartType,
-                esql: request.panelInput.esql,
-              }),
-            }))
-          ),
-        ] as const
-    )
-  );
-
-  return new Map(resolvedRequestsByOperationIndex);
-};
-
-/**
- * Return the resolved create results for one operation during the apply phase.
- * Throw if an operation that should have resolved create results does not have them.
- */
-const getResolvedVisualizationCreationRequests = ({
-  resolvedRequestsByOperationIndex,
-  operationIndex,
-  operationType,
-}: {
-  resolvedRequestsByOperationIndex: Map<number, ResolvedVisualizationCreationRequest[]>;
-  operationIndex: number;
-  operationType: VisualizationCreationOperationType;
-}): ResolvedVisualizationCreationRequest[] => {
-  const resolvedRequests = resolvedRequestsByOperationIndex.get(operationIndex);
-
-  if (!resolvedRequests) {
-    throw new Error(
-      `Missing pre-resolved visualization requests for ${operationType} operation at index ${operationIndex}.`
-    );
-  }
-
-  return resolvedRequests;
-};
-
-/**
- * Turn resolved create results into dashboard panels and append any failures.
- * Successful panels are kept even when sibling requests fail.
- */
-const materializeResolvedVisualizationPanels = ({
-  resolvedRequests,
-  failures,
-}: {
-  resolvedRequests: ResolvedVisualizationCreationRequest[];
-  failures: VisualizationFailure[];
-}): Array<{ request: VisualizationCreationRequest; panel: AttachmentPanel }> => {
-  const successfulPanels: Array<{ request: VisualizationCreationRequest; panel: AttachmentPanel }> =
-    [];
-
-  for (const { request, resolvedPanel } of resolvedRequests) {
-    if (resolvedPanel.type === 'failure') {
-      failures.push(resolvedPanel.failure);
-      continue;
-    }
-
-    successfulPanels.push({
-      request,
-      panel: {
-        id: uuidv4(),
-        type: resolvedPanel.visContent.type,
-        config: resolvedPanel.visContent.config,
-        grid: request.panelInput.grid,
-      },
-    });
-  }
-
-  return successfulPanels;
-};
+export type VisualizationPanelInput = z.infer<typeof visualizationPanelInputSchema>;
+export type CreateVisualizationPanelInput = z.infer<typeof createVisualizationPanelSchema>;
 
 const createEmptyDashboardData = (): DashboardAttachmentData => ({
   title: 'User Dashboard',
@@ -408,278 +233,21 @@ export const executeDashboardOperations = async ({
     requestsByOperationIndex: visualizationCreationRequests,
     resolveVisualizationConfig,
   });
+  const context: OperationExecutionContext = {
+    logger,
+    failures,
+    resolvedVisualizationCreationRequests,
+    resolvePanelsFromAttachments,
+    resolveVisualizationConfig,
+  };
 
   for (const [operationIndex, operation] of operations.entries()) {
-    switch (operation.operation) {
-      case 'set_metadata': {
-        if (operation.title === undefined && operation.description === undefined) {
-          logger.debug('Skipping empty set_metadata operation');
-          break;
-        }
-
-        const metadataPatch = {
-          ...(operation.title !== undefined ? { title: operation.title } : {}),
-          ...(operation.description !== undefined ? { description: operation.description } : {}),
-        };
-        nextDashboardData = {
-          ...nextDashboardData,
-          ...metadataPatch,
-        };
-        break;
-      }
-
-      case 'add_markdown': {
-        const markdownPanel = {
-          type: MARKDOWN_EMBEDDABLE_TYPE,
-          config: { content: operation.markdownContent },
-          grid: operation.grid,
-          id: uuidv4(),
-        };
-        nextDashboardData = appendPanelsToDashboard({
-          dashboardData: nextDashboardData,
-          panelsToAdd: [markdownPanel],
-          sectionId: operation.sectionId,
-        });
-
-        break;
-      }
-
-      case 'add_panels_from_attachments': {
-        for (const item of operation.items) {
-          const result = resolvePanelsFromAttachments([
-            {
-              attachmentId: item.attachmentId,
-              grid: item.grid,
-            },
-          ]);
-          if (result.panels.length > 0) {
-            nextDashboardData = appendPanelsToDashboard({
-              dashboardData: nextDashboardData,
-              panelsToAdd: result.panels,
-              sectionId: item.sectionId,
-            });
-          }
-          failures.push(...result.failures);
-        }
-        break;
-      }
-
-      case 'create_visualization_panels': {
-        const panelsToAdd = materializeResolvedVisualizationPanels({
-          resolvedRequests: getResolvedVisualizationCreationRequests({
-            resolvedRequestsByOperationIndex: resolvedVisualizationCreationRequests,
-            operationIndex,
-            operationType: operation.operation,
-          }),
-          failures,
-        });
-
-        for (const { request, panel } of panelsToAdd) {
-          nextDashboardData = appendPanelsToDashboard({
-            dashboardData: nextDashboardData,
-            panelsToAdd: [panel],
-            sectionId:
-              request.operationType === 'create_visualization_panels'
-                ? request.sectionId
-                : undefined,
-          });
-        }
-        break;
-      }
-
-      case 'edit_visualization_panels': {
-        if (!resolveVisualizationConfig) {
-          throw new Error(
-            'Inline visualization resolver is required for edit_visualization_panels operations.'
-          );
-        }
-
-        for (const panelInput of operation.panels) {
-          const existingPanel = findPanelById(nextDashboardData.panels, panelInput.panelId);
-          if (!existingPanel) {
-            const failureResult = createVisualizationFailureResult(
-              operation.operation,
-              panelInput.panelId,
-              `Panel "${panelInput.panelId}" not found.`
-            );
-            if (failureResult.type === 'failure') {
-              failures.push(failureResult.failure);
-            }
-            continue;
-          }
-          const resolvedPanel = await resolveVisualizationConfig({
-            operationType: operation.operation,
-            identifier: panelInput.panelId,
-            nlQuery: panelInput.query,
-            chartType: panelInput.chartType,
-            esql: panelInput.esql,
-            existingPanel,
-          });
-
-          if (resolvedPanel.type === 'failure') {
-            failures.push(resolvedPanel.failure);
-            continue;
-          }
-
-          const updateResult = updatePanelInDashboard({
-            dashboardData: nextDashboardData,
-            panelId: panelInput.panelId,
-            transformPanel: (panel) => ({
-              ...{
-                ...panel,
-                ...resolvedPanel.visContent,
-              },
-            }),
-          });
-
-          if (!updateResult.updated) {
-            failures.push({
-              type: 'edit_visualization_panels',
-              identifier: panelInput.panelId,
-              error: `Panel "${panelInput.panelId}" not found.`,
-            });
-            continue;
-          }
-
-          nextDashboardData = updateResult.dashboardData;
-        }
-        break;
-      }
-
-      case 'update_panel_layouts': {
-        const recordMissingPanelFailure = (panelId: string) => {
-          failures.push({
-            type: 'update_panel_layouts',
-            identifier: panelId,
-            error: `Panel "${panelId}" not found.`,
-          });
-        };
-
-        for (const { panelId, grid, sectionId } of operation.panels) {
-          // sectionId omitted: do not move the panel
-          if (sectionId === undefined) {
-            const updateResult = updatePanelInDashboard({
-              dashboardData: nextDashboardData,
-              panelId,
-              transformPanel: (panel) => ({
-                ...panel,
-                ...(grid ? { grid } : {}),
-              }),
-            });
-
-            if (!updateResult.updated) {
-              recordMissingPanelFailure(panelId);
-              continue;
-            }
-
-            nextDashboardData = updateResult.dashboardData;
-            continue;
-          }
-          // sectionId provided: move the panel to that section, or to the top level when null
-          const removalResult = removePanelsFromDashboard({
-            dashboardData: nextDashboardData,
-            panelIdsToRemove: [panelId],
-          });
-          const { dashboardData: dashboardAfterRemoval, removedPanels } = removalResult;
-
-          if (removedPanels.length === 0) {
-            recordMissingPanelFailure(panelId);
-            continue;
-          }
-
-          const [panelToMove] = removedPanels;
-          nextDashboardData = appendPanelsToDashboard({
-            dashboardData: dashboardAfterRemoval,
-            panelsToAdd: [
-              {
-                ...panelToMove,
-                ...(grid ? { grid } : {}),
-              },
-            ],
-            // sectionId targets a section; null promotes the panel to the top level
-            sectionId: sectionId ?? undefined,
-          });
-        }
-        break;
-      }
-
-      case 'add_section': {
-        let nextSection: DashboardSection = {
-          id: uuidv4(),
-          title: operation.title,
-          collapsed: false,
-          grid: operation.grid,
-          panels: [],
-        };
-
-        if (operation.panels) {
-          const sectionPanels = materializeResolvedVisualizationPanels({
-            resolvedRequests: getResolvedVisualizationCreationRequests({
-              resolvedRequestsByOperationIndex: resolvedVisualizationCreationRequests,
-              operationIndex,
-              operationType: operation.operation,
-            }),
-            failures,
-          }).map(({ panel }) => panel);
-
-          nextSection = {
-            ...nextSection,
-            panels: sectionPanels,
-          };
-        }
-
-        nextDashboardData = {
-          ...nextDashboardData,
-          panels: [...nextDashboardData.panels, nextSection],
-        };
-        break;
-      }
-
-      case 'remove_section': {
-        const sectionIndex = findSectionIndex(nextDashboardData.panels, operation.id);
-        if (sectionIndex === -1) {
-          throw new Error(`Section "${operation.id}" not found.`);
-        }
-
-        const sectionToRemove = nextDashboardData.panels[sectionIndex] as DashboardSection;
-        const nextPanels = nextDashboardData.panels.filter((_, i) => i !== sectionIndex);
-
-        if (operation.panelAction === 'delete') {
-          nextDashboardData = {
-            ...nextDashboardData,
-            panels: nextPanels,
-          };
-          break;
-        }
-
-        const baseY = getWidgetsBottomY(nextPanels);
-        const promotedPanels = sectionToRemove.panels.map((panel) => ({
-          ...panel,
-          grid: {
-            ...panel.grid,
-            y: baseY + panel.grid.y,
-          },
-        }));
-
-        nextDashboardData = {
-          ...nextDashboardData,
-          panels: [...nextPanels, ...promotedPanels],
-        };
-        break;
-      }
-
-      case 'remove_panels': {
-        const { dashboardData: dashboardWithoutPanels, removedPanels } = removePanelsFromDashboard({
-          dashboardData: nextDashboardData,
-          panelIdsToRemove: operation.panelIds,
-        });
-        if (removedPanels.length > 0) {
-          nextDashboardData = dashboardWithoutPanels;
-          logger.debug(`Removed ${removedPanels.length} panels from dashboard`);
-        }
-        break;
-      }
-    }
+    nextDashboardData = await executeOperationHandler({
+      dashboardData: nextDashboardData,
+      operation,
+      operationIndex,
+      context,
+    });
   }
 
   return {
